ai/api-server/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
boto3==1.34.80
botocore==1.34.80
click==8.1.7
colorama==0.4.6
fastapi==0.110.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
jmespath==1.0.1
mysqlclient==2.2.4
pika==1.3.2
py-redis==1.1.1
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
redis==5.0.3
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
